Move text highlighting test to MediaLayout from QuestionWidget
This is the actual component shared between all uses of this feature
