+++ collect_app/src/test/java/org/odk/collect/android/support/RobolectricHelpers.java
@@ -1,9 +1,13 @@
 package org.odk.collect.android.support;
 
+import androidx.fragment.app.FragmentActivity;
+
+import org.odk.collect.android.R;
 import org.odk.collect.android.application.Collect;
 import org.odk.collect.android.injection.config.AppDependencyComponent;
 import org.odk.collect.android.injection.config.AppDependencyModule;
 import org.odk.collect.android.injection.config.DaggerAppDependencyComponent;
+import org.robolectric.Robolectric;
 import org.robolectric.RuntimeEnvironment;
 
 public class RobolectricHelpers {
@@ -20,5 +24,12 @@ public class RobolectricHelpers {
 
     public static AppDependencyComponent getApplicationComponent() {
         return ((Collect) RuntimeEnvironment.application).getComponent();
+    }
+
+    public static FragmentActivity createThemedActivity() {
+        FragmentActivity activity = Robolectric.setupActivity(FragmentActivity.class);
+        activity.setTheme(R.style.LightAppTheme); // Needed so attrs are available
+
+        return activity;
     }
 }
